Dialzyer, handle lexer error line numbers, test to its own file
