+++ lib/thrift/exceptions.ex
@@ -30,6 +30,8 @@ defmodule Thrift.FileParseException do
 
   defexception message: nil
 
+  @doc false  # Exception callback, should not be called by end user
+  @spec exception({Thrift.Parser.FileRef.t, term}) :: Exception.t
   def exception({file_ref, error}) do
     msg = "Error parsing thrift file #{file_ref.path} #{format_error(error)}"
     %Thrift.FileParseException{message: msg}
@@ -37,6 +39,9 @@ defmodule Thrift.FileParseException do
 
   # display the line number if we get it
   defp format_error({line_no, :thrift_parser, errors}) do
+    "on line #{line_no}: #{inspect errors}"
+  end
+  defp format_error({{line_no, :thrift_lexer, errors}, _}) do
     "on line #{line_no}: #{inspect errors}"
   end
   defp format_error(error) do
