+++ test/parser_test.exs
@@ -1,9 +1,8 @@
 defmodule ParserTest do
-  @project_root Path.expand("../", __DIR__)
 
   use ExUnit.Case, async: true
 
-  import Thrift.Parser, only: [parse: 1, parse: 2, parse_file: 1]
+  import Thrift.Parser, only: [parse: 1, parse: 2]
 
   alias Thrift.Parser.Models.Constant
   alias Thrift.Parser.Models.Exception
@@ -697,45 +696,5 @@ defmodule ParserTest do
 
     File.read!("./test/fixtures/app/thrift/shared.thrift")
     |> parse
-  end
-
-  test "parsing a file with syntax errors" do
-    contents = """
-    stract Typo {
-        1: optional i32 id
-    }
-    """
-
-    assert {:error, _} = parse(contents)
-
-    dir = Path.join([@project_root, "tmp", "syntax_error_test"])
-    File.mkdir_p!(dir)
-    path = Path.join(dir, "syntax_error.thrift")
-    File.write!(path, contents)
-
-    assert_raise(
-      Thrift.FileParseException,
-      ~r/#{path} on line 1:/,
-      fn -> parse_file(path) end
-    )
-
-    other_path = Path.join(dir, "includes_syntax_error.thrift")
-    File.write!(other_path, """
-    include "syntax_error.thrift"
-
-    struct NoError {
-      1: optional i32 id
-    }
-    """)
-
-    # should raise an error on the included file,
-    # since that is where the syntax error is
-    assert_raise(
-      Thrift.FileParseException,
-      ~r/#{path} on line 1:/,
-      fn -> parse_file(other_path) end
-    )
-
-    File.rm_rf!(dir)
   end
 end
